✨ Move buy-required action into Process component
What: move Buy required items UI/logic from process detail page into

the shared Process component so embeds inherit behavior.

Why: embedded processes (quests/items) must expose purchasing of missing

requirements just like the detail route.

How to test: npx vitest run frontend/src/pages/process/[slug]/__tests__/ProcessView.spec.ts frontend/src/components/__tests__/Process.spec.ts
